requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
Pillow==9.2.0
easy-thumbnails==2.8.1
djangorestframework==3.14.0
pyyaml==6.0.1
drf-yasg==1.21.7
django-rest-swagger==2.2.0
django-cors-headers==4.3.0
tk==0.1.0
djangorestframework-simplejwt==5.3.0
markdown==3.5.1
humanize==4.9.0